+++ clients/python/balancing_services/models/__init__.py
@@ -14,10 +14,29 @@ from .balancing_capacity_prices_response import BalancingCapacityPricesResponse
 from .balancing_capacity_volume import BalancingCapacityVolume
 from .balancing_capacity_volumes import BalancingCapacityVolumes
 from .balancing_capacity_volumes_response import BalancingCapacityVolumesResponse
+from .balancing_energy_activated_volume_history import (
+    BalancingEnergyActivatedVolumeHistory,
+)
+from .balancing_energy_activated_volume_history_response import (
+    BalancingEnergyActivatedVolumeHistoryResponse,
+)
+from .balancing_energy_activated_volume_revision import (
+    BalancingEnergyActivatedVolumeRevision,
+)
 from .balancing_energy_bid_period import BalancingEnergyBidPeriod
 from .balancing_energy_bids import BalancingEnergyBids
 from .balancing_energy_bids_response import BalancingEnergyBidsResponse
+from .balancing_energy_offered_volume_history import BalancingEnergyOfferedVolumeHistory
+from .balancing_energy_offered_volume_history_response import (
+    BalancingEnergyOfferedVolumeHistoryResponse,
+)
+from .balancing_energy_offered_volume_revision import (
+    BalancingEnergyOfferedVolumeRevision,
+)
 from .balancing_energy_price import BalancingEnergyPrice
+from .balancing_energy_price_history import BalancingEnergyPriceHistory
+from .balancing_energy_price_history_response import BalancingEnergyPriceHistoryResponse
+from .balancing_energy_price_revision import BalancingEnergyPriceRevision
 from .balancing_energy_prices import BalancingEnergyPrices
 from .balancing_energy_prices_response import BalancingEnergyPricesResponse
 from .balancing_energy_volume import BalancingEnergyVolume
@@ -85,10 +104,19 @@ __all__ = (
     "BalancingCapacityVolume",
     "BalancingCapacityVolumes",
     "BalancingCapacityVolumesResponse",
+    "BalancingEnergyActivatedVolumeHistory",
+    "BalancingEnergyActivatedVolumeHistoryResponse",
+    "BalancingEnergyActivatedVolumeRevision",
     "BalancingEnergyBidPeriod",
     "BalancingEnergyBids",
     "BalancingEnergyBidsResponse",
+    "BalancingEnergyOfferedVolumeHistory",
+    "BalancingEnergyOfferedVolumeHistoryResponse",
+    "BalancingEnergyOfferedVolumeRevision",
     "BalancingEnergyPrice",
+    "BalancingEnergyPriceHistory",
+    "BalancingEnergyPriceHistoryResponse",
+    "BalancingEnergyPriceRevision",
     "BalancingEnergyPrices",
     "BalancingEnergyPricesResponse",
     "BalancingEnergyVolume",